requirements.txt
blinker==1.6.3
build==1.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.13.1
elasticsearch==8.14.0
filelock==3.15.4
Flask==3.0.0
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pillow==10.4.0
pip-tools==7.3.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==72.1.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.1
wheel==0.41.3
